artisan-platform/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
alembic
flask
flask-cors
psycopg2-binary
jose
passlib
pydantic-settings
flask_sqlalchemy
mysql-connector-python
python-jose